# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Purchase Approval Flow API
  version: 1.0.0
  description: |
    Serverless purchase-approval workflow with CONSENT SIGNATURE and immutable
    evidence. Flow: create request → unique link per approver → OTP →
    approve/reject with CAS → PDF evidence on completion.

    This document is the authoritative 12-endpoint contract for the reviewer
    driving the demo. It mirrors `openspec/changes/purchase-approval-flow/design-api.md`
    (the planning source of truth) and pins the error→HTTP policy plus example
    curl flows. The system is NOT a payment system and NOT PKI/cryptographic
    signing — the "signature" is registered consent evidence (registered snapshot
    name + timestamp after OTP validation).

    Base path: every route is served under the API Gateway stage (e.g. `/dev`).
servers:
  - url: http://localhost:4000/dev
    description: Local development (serverless-offline)
  - url: https://<api-id>.execute-api.<region>.amazonaws.com/dev
    description: Deployed API Gateway (replace placeholders after `sls deploy`)
...
paths:
...
  /api/purchase-requests/{requestId}/approvers/{email}/recover:
    post:
      tags: [requests]
      summary: 'Re-send OTP for a LOCKED approver (requester-initiated recovery)'
      description: |
        Requester-initiated recovery of a LOCKED approver's OTP (DECISIONS #25).
        `locked` is TRUE only when the approver's token is `INVALIDATED_LOCKOUT`
        (3 failed OTP attempts). This resets the locked approver to ACTIVE,
        clears the failed-attempt count and issues them a fresh OTP (mailed).
        It is SCOPED to a LOCKED approver ONLY — an innocent pending approver's
        OTP is never re-issued by an action they don't control (their OTP only
        changes via their OWN issue/regenerate flow).
      operationId: recoverApproverOtp
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: string
          description: Purchase request id
        - in: path
          name: email
          required: true
          schema:
            type: string
            format: email
          description: Approver email
      responses:
        '201':
          description: Locked approver recovered with a fresh 3-minute OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpIssueResponse'
        '404':
          description: Unknown request or approver not part of the request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Approver is NOT locked (recovery is locked-only) — no OTP re-issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '410':
          description: Request is terminal (COMPLETED/REJECTED) — no recovery
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/approvals/{requestId}/token/{token}/otp:
    post:
      tags: [approver-otp]
      summary: 'Issue OTP (#7)'
      description: |
        Resolves the token and issues a 6-digit OTP (mailed via the mock outbox,
        disclosed in /mock-mail), valid 180 seconds. Gate chain:
        request terminal → 410 · approver locked → 403 · already acted → 409 ·
        unknown request/token → 404. The approver is identified by the token —
        no request body is read.
      operationId: issueOtp
      parameters:
        - $ref: '#/components/parameters/ApprovalRequestId'
        - $ref: '#/components/parameters/ApproverToken'
      responses:
        '201':
          description: OTP issued and mailed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpIssueResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '403':
          $ref: '#/components/responses/LockoutError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '410':
          $ref: '#/components/responses/TerminalError'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/approvals/{requestId}/token/{token}/otp/validate:
    post:
      tags: [approver-otp]
      summary: 'Validate OTP (#8)'
      description: |
        Consumes the OTP with an atomic conditional delete — exactly ONE
        concurrent submission of the correct code wins; losers get 410
        (ExpiredOtpError). A successful validation durably writes a
        `validatedAt` marker on the approver row, required by #10/#11 (missing
        → 401). Wrong code → 401 with `attemptsRemaining`; 3 failures atomically
        lock the approver (403). Expiry is enforced in code.
      operationId: validateOtp
      parameters:
        - $ref: '#/components/parameters/ApprovalRequestId'
        - $ref: '#/components/parameters/ApproverToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  description: 6-digit numeric code
                  pattern: '^[0-9]{6}$'
                  example: '123456'
      responses:
        '200':
          description: OTP valid and consumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpValidateResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Wrong OTP code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WrongOtpError'
        '403':
          $ref: '#/components/responses/LockoutError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '410':
          description: OTP expired or already consumed (single-use) — regenerate is allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/approvals/{requestId}/token/{token}/otp/regenerate:
    post:
      tags: [approver-otp]
      summary: 'Regenerate OTP (#9)'
      description: |
        Issues a fresh OTP + 180s TTL and resets failed attempts — only while the
        approver token is still ACTIVE (locked → 403, already acted → 409,
        terminal request → 410). No request body is read.
      operationId: regenerateOtp
      parameters:
        - $ref: '#/components/parameters/ApprovalRequestId'
        - $ref: '#/components/parameters/ApproverToken'
      responses:
        '201':
          description: New OTP issued and mailed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpIssueResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '403':
          $ref: '#/components/responses/LockoutError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '410':
          $ref: '#/components/responses/TerminalError'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  parameters:
...
    ApprovalRequestId:
      description: Purchase request id (UUID) of the approval
      in: path
      required: true
      name: requestId
      schema:
        type: string
        format: uuid
    ApproverToken:
      description: Unique per-approver token from the approval link
      in: path
      required: true
      name: token
      schema:
        type: string
        format: uuid
  schemas:
...
    OtpIssueResponse:
      type: object
      required: [expiresInSeconds]
      properties:
        expiresInSeconds:
          type: integer
          example: 180
    OtpValidateResponse:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
          const: true
    WrongOtpError:
      type: object
      required: [error, message, attemptsRemaining]
      properties:
        error:
          type: string
          example: WrongOtpError
        message:
          type: string
        attemptsRemaining:
          type: integer
          example: 2
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Domain error class name, e.g. UnknownRequestError, TerminalRequestError
        message:
          type: string
  responses:
    ValidationError:
      description: Validation failed (empty fields, bad email, amount ≤ 0 or > 2 decimals, wrong approver count, malformed JSON)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: InvalidEmailError
            message: Email must be a valid address
    ConflictError:
      description: Duplicate email (register) or already acted (approve/reject duplicate, OTP on acted token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: AlreadyActedError
            message: This approver already signed the request
    NotFoundError:
      description: Unknown request id, unknown token, or unknown registry email on create
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: UnknownRequestError
            message: Request req-1 not found
    LockoutError:
      description: Approver is locked out (3 failed OTP attempts) or token not ACTIVE
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: LockedOutError
            message: This approver is locked out
    TerminalError:
      description: Request is in a terminal global state (COMPLETED/REJECTED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: TerminalRequestError
            message: Request req-1 is already COMPLETED; no OTP flow is offered
    ServerError:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...